using injected-script to register all frames in the page and load data
inspectedWindow.eval can accept a frameURL, but we need to get a list of
frames by having the injected-script register each frame it is injected
into. Then $0 can be passed to functions in the injected script in order
read the data from the selected element in whatever frame it is in.